Allow custom PATH variables. Expose bindgen as a feature.
AMD have documented that HIP_PATH is to be set when a custom installation path
is used; the relevant build.rs file now searches and uses this variable.

Unfortunately there doesn't appear to be an equivalent for hipblas. I've taken
the liberty of searching for HIP_BLAS_PATH and HIPBLAS_PATH, failing those,
assuming a path based off of HIP_PATH.

bindgen is now a feature in hip-sys, which triggers the bindgen features of
hip-runtime-sys and hipblas-sys.